media handler

+++ core/internal/models/media.go
@@ -11,7 +11,7 @@ type Media struct {
 	URL          string    `gorm:"type:varchar(255);not null"`
 	FileName     string    `gorm:"type:varchar(255);not null"`
 	FileType     string    `gorm:"type:varchar(50);not null"`
-	FileSize     int       `gorm:"not null"`
+	FileSize     int64     `gorm:"not null"`
 	UploadTime   time.Time `gorm:"not null"`
 	Description  string    `gorm:"type:text"`
 	BucketName   string    `gorm:"type:varchar(255);not null"`
